requirements.txt
pydantic>=2.0
pytest>=7.0
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
supabase>=2.0.0
python-dotenv>=1.0.0
